Feature/cicd integration (#7)
COMMENTS:
There are multiple bugfixes in this branch to allow the ci build to succeed:

The app module needed a way to be closed gracefully. With it, the Redis connection is also closed. This behavior was making mocha to hang.
On the test files, I have removed the server utilities and use the withEnvironment wrapper as a default option. The test passed locally but not on the CI environment because the .env variables were loading on a local environment but not on CI. I should not load the Env module to avoid these errors.
I add a try/finally handler to close the server whenever the tests failed. This behavior has been included in withEnvironment wrapper and was the root cause for mocha to hang on failed tests. In the CI environment, the build was running for 10 minutes before failure.

UPDATE:
4. On the test code, all environment variables are strings.
5. The Env module is no longer loaded on withEnvironment wrapper to isolate local and CI environments.

UPDATE 2:
The Redis database was being reused and some tests broke when the environment variables were strings again. In this pull request I added an option to flush the Redis database on close. I also added the OVERALL_REQUESTS_DURATION environment variables to the tests that requested it.

COMMITS:

* Add travis ci config file

* Add close function to app module in order to close the redis connection and prevent mocha from hanging when the test finishes

* Add redis to travis configuration

* Split rate limiter headers tests

* Uninstall unused dependencies

* Debug on CI server

* Debug sent headers on CI environment

* Add coverage script

* Add exit argument on test script

* Fix test environment

* Remove test environment variables when they were not necessary

* Remove Env module loading to avoid differences between local and CI environment

* All environment variables passed to app converted to string to avoid false positives

* Add an option to flush the redis database on close

* Add OVERALL_REQUESTS_DURATION to test environment variables